mySQL Database
Created mySQL Database to hold brew info.
Made home page to display brews that are ready to drink.
Made a schedule page that displays what needs to be done and when.
Button in table allows quick updating of database with minimal steps.
Made a details page that shows an overview of a specific brew and
ability to update database from that page (not set up right now).